Back-end and front-end separated into separate folders
