+++ web-api/.vscode/tasks.json
@@ -1,0 +1,15 @@
+{
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "build",
+            "command": "dotnet",
+            "type": "process",
+            "args": [
+                "build",
+                "${workspaceFolder}/SafeHouse/SafeHouse/SafeHouse.csproj"
+            ],
+            "problemMatcher": "$msCompile"
+        }
+    ]
+}
