+++ web-api/SafeHouse/SafeHouse.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2037
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SafeHouse", "SafeHouse\SafeHouse.csproj", "{5007D3BC-D27A-4BB5-BA9D-491E5660C512}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5007D3BC-D27A-4BB5-BA9D-491E5660C512}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5007D3BC-D27A-4BB5-BA9D-491E5660C512}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5007D3BC-D27A-4BB5-BA9D-491E5660C512}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5007D3BC-D27A-4BB5-BA9D-491E5660C512}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EBFF5DBA-B471-4CC3-A215-055662FFD84D}
+	EndGlobalSection
+EndGlobal
