+++ web-api/SafeHouse/SafeHouse/appsettings.json
@@ -1,0 +1,18 @@
+﻿{
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=13.93.28.210;Database=VegaIT.Svratiste.Development;persist security info=True;user id=svratiste.dev;password=e(NW#4Ys\"jc4F@5n"
+  },
+  "Logging": {
+    "IncludeScopes": false,
+    "Debug": {
+      "LogLevel": {
+        "Default": "Warning"
+      }
+    },
+    "Console": {
+      "LogLevel": {
+        "Default": "Warning"
+      }
+    }
+  }
+}
