+++ web-client/SafeHouseClient/public/favicon.ico
