+++ web-client/SafeHouseClient/public/manifest.json
