+++ web-client/SafeHouseClient/src/registerServiceWorker.js
